Created Bookshelf, Created Books, Created Triggers, Created CorrectBook Script
Rotated the camera to a birds-eye view
Created an outline of a bookshelf using game objects.
Created cubes to resemble books, added text to assign each book a letter.
Added Trigger Points under the books with the correct letter
Created OnTrigger script for the book to turn green when in the correct position.